rearranged group related integration tests
